src/asammdf/plugins/aiaspro/requirements.txt
# AIASPRO Plugin Requirements
pydantic-ai>=0.0.7
openai>=1.0.0
scikit-learn>=1.3.0
scipy>=1.10.0
cryptography>=3.4.0
